requirements.txt
biopython
biosimulators_bionetgen
biosimulators_utils[bngl]
pyyaml
requests
